+++ src/Runtime/Animation/mmdRuntimeModelAnimationContainer.ts
@@ -401,6 +401,7 @@ declare module "../../Loader/Animation/mmdModelAnimationContainer" {
 }
 
 /**
+ * @internal
  * Create runtime model animation
  * @param model Bind target
  * @param retargetingMap Animation bone name to model bone name map