+++ src/Test/Scene/gltfBpmxTestScene.ts
@@ -160,8 +160,8 @@ export class SceneBuilder implements ISceneBuilder {
         mmdRuntime.setManualAnimationDuration(mmdAnimation.endFrame);
 
         mmdRuntime.addAnimatable(mmdCamera);
-        mmdCamera.addAnimation(mmdAnimation);
-        mmdCamera.setAnimation("motion");
+        const cameraAnimationHandle = mmdCamera.createRuntimeAnimation(mmdAnimation);
+        mmdCamera.setRuntimeAnimation(cameraAnimationHandle);
 
         mmdMesh.scaling.scaleInPlace(14.3);
         mmdMesh.scaling.z *= -1;
@@ -247,8 +247,8 @@ export class SceneBuilder implements ISceneBuilder {
             }
         );
         mmdModel.morph.setMorphWeight("口_真顔", 0.2);
-        mmdModel.addAnimation(mmdAnimation);
-        mmdModel.setAnimation("motion");
+        const modelRuntimeAnimationHandle = mmdModel.createRuntimeAnimation(mmdAnimation);
+        mmdModel.setRuntimeAnimation(modelRuntimeAnimationHandle);
 
         const translationMatrix = mmdMesh.getWorldMatrix().clone();
         translationMatrix.removeRotationAndScaling();