+++ src/Test/Scene/webXrTestScene.ts
@@ -140,8 +140,8 @@ export class SceneBuilder implements ISceneBuilder {
         mmdRuntime.setManualAnimationDuration(mmdAnimation.endFrame);
 
         mmdRuntime.addAnimatable(mmdCamera);
-        mmdCamera.addAnimation(mmdAnimation);
-        mmdCamera.setAnimation("motion");
+        const cameraAnimationHandle = mmdCamera.createRuntimeAnimation(mmdAnimation);
+        mmdCamera.setRuntimeAnimation(cameraAnimationHandle);
 
         {
             for (const mesh of modelMesh.metadata.meshes) {
@@ -153,8 +153,8 @@ export class SceneBuilder implements ISceneBuilder {
             const mmdModel = mmdRuntime.createMmdModel(modelMesh, {
                 buildPhysics: true
             });
-            mmdModel.addAnimation(mmdAnimation);
-            mmdModel.setAnimation("motion");
+            const modelRuntimeAnimationHandle = mmdModel.createRuntimeAnimation(mmdAnimation);
+            mmdModel.setRuntimeAnimation(modelRuntimeAnimationHandle);
 
             AttachToBone(scene, mmdModel, {
                 directionalLightPosition: directionalLight.position,